# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.DashboardStats:
    properties:
      compliance:
        description: allowed, denied, unknown
        items:
          $ref: '#/definitions/handlers.DataPoint'
        type: array
      ecosystems:
        items:
          $ref: '#/definitions/handlers.DataPoint'
        type: array
      licenses:
        items:
          $ref: '#/definitions/handlers.DataPoint'
        type: array
      total_dependencies:
        type: integer
      vulnerabilities:
        items:
          $ref: '#/definitions/handlers.DataPoint'
        type: array
    type: object
  handlers.DataPoint:
    properties:
      name:
        type: string
      value:
        type: integer
    type: object
  handlers.LicenseCountResponse:
    properties:
      count:
        type: integer
      name:
        type: string
    type: object
  handlers.LicenseUsageResponse:
    properties:
      component_name:
        type: string
      product_name:
        type: string
      repository_name:
        type: string
      version:
        type: string
    type: object
  handlers.PolicyBreach:
    properties:
      component_name:
        type: string
      license:
        type: string
      product_name:
        type: string
      reason:
        type: string
      repository_name:
        type: string
    type: object
  handlers.PolicyDetailsResponse:
    properties:
      breaches:
        items:
          $ref: '#/definitions/handlers.PolicyBreach'
        type: array
      compliance_percent:
        type: number
      denied_components:
        type: integer
      policy: {}
      products: {}
      total_components:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Scapegoat SBOM and Vulnerability Management API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scapegoat API
  version: "1.0"
paths:
  /dashboard/stats:
    get:
      description: Get aggregated stats including total dependencies, license distribution,
        ecosystem distribution, compliance status, and vulnerability severity counts.
      parameters:
      - description: Organisation ID to filter by
        in: query
        name: organisation_id
        type: integer
      - description: Product ID to filter by
        in: query
        name: product_id
        type: integer
      - description: Repository ID to filter by
        in: query
        name: repository_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DashboardStats'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get dashboard statistics
      tags:
      - dashboard
  /licenses:
    get:
      description: Get a list of all unique licenses found in the latest SBOMs, with
        counts.
      parameters:
      - description: Organisation ID to filter by
        in: query
        name: organisation_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.LicenseCountResponse'
            type: array
      summary: List all licenses
      tags:
      - compliance
  /licenses/{name}:
    get:
      description: Get a list of components across all repositories that use a specific
        license.
      parameters:
      - description: License Name (use 'Unknown' for empty licenses)
        in: path
        name: name
        required: true
        type: string
      - description: Organisation ID to filter by
        in: query
        name: organisation_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.LicenseUsageResponse'
            type: array
      summary: Get license details
      tags:
      - compliance
...
  /policies/{id}:
    get:
      description: Get detailed information about a policy, including products using
        it, compliance percentage, and a list of components breaching the policy.
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PolicyDetailsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get policy details
      tags:
      - policies
...
swagger: "2.0"